SceneAlgo : Move applyCameraGlobals and render adaptors from RendererAlgo
These are the only public parts of RendererAlgo
